Add 1st working iteration for the problem
This is actually a simple solution. It produces a list of tuples,
one tuple for each possible input value. Then, when the 'signal'
is received, 'solver_par' puts the list in the 'box'.

A possible adjustment would be to add an extra functionality that
allows the computation of the list to stop (if not finished already)
as soon as the signal is received and return only the list of tuples
that are computed till this point.